NOJIRA forenkle kodeverk-oppsett
